Require Jenkins 2.289.1 (recommended minimum version)
https://www.jenkins.io/doc/developer/plugin-development/choosing-jenkins-baseline/
recommends Jenkins 2.289.1 as the minimum version.  It has the following
benefits:

* New enough that recent changes like the replacement of JSR-305
  annotations with spotbugs annotations is visible
* New enough that the UI modernization (table to div) is always
  available
* Less out of date than Jenkins 2.200

In addition to those reasons, the installation data at
https://stats.jenkins.io/pluginversions/nuclei.html shows that only 3
installations are running a version older than 2.289.1.  Those three
installations are probably not going to upgrade to a new plugin version
unless they also upgrade to a new Jenkins version.

In order to update to the 2.289.1 minimum version, the JSR-305
annotations have been replaced with spotbugs annotations.